Emit MPRIS PropertiesChanged on shuffle/loop toggle; fix format warnings
Shuffle and LoopStatus now emit PropertiesChanged when toggled via GUI
buttons so connected status bars update immediately. Also fixes format
type mismatches and a truncation warning in library.cpp, and a
sign-conversion warning in queue.cpp.

+++ src/gui/views/queue.cpp
@@ -82,7 +82,7 @@ void draw_queue_view(AppState& st) {
         char label[512];
         std::snprintf(label, sizeof(label), "%d  %s",
                       i + 1,
-                      snap_queue[i].stem().string().c_str());
+                      snap_queue[static_cast<std::size_t>(i)].stem().string().c_str());
 
         if (ImGui::Selectable(label, current)) {
             st.queue_jump_to(i);
